refactor(core): move command contracts to neutral module
