+++ src/process/command.ts
@@ -1,0 +1,21 @@
+export type CommandResult = {
+  code: number;
+  stdout: string;
+  stderr: string;
+};
+
+export type CommandRunOptions = {
+  cwd?: string;
+  env?: Record<string, string | undefined>;
+  onStdout?: (chunk: string) => void;
+  onStderr?: (chunk: string) => void;
+  signal?: AbortSignal;
+};
+
+export type CommandRunner = {
+  run(
+    command: string,
+    args: string[],
+    options?: CommandRunOptions,
+  ): Promise<CommandResult>;
+};
